requirements.txt
numpy
matplotlib
perlin_noise
pillow
trimesh
pyglet<2